Move NandVector stuff from syntax into evaluator.py
