+++ nand/evaluator.py
@@ -445,3 +445,129 @@ def custom_op(f):
 
     # Wrap the function with a semaphore value to indicate it's not Nand
     return (None, f)
+
+
+class NandVectorWrapper:
+    """Convenient syntax around a NandVector. You get one of these from run(chip).
+    """
+    
+    def __init__(self, vector):
+        self._vector = vector
+        
+    def __setattr__(self, name, value):
+        """Set the value of a single- or multiple-bit input."""
+        if name.startswith('_'): return object.__setattr__(self, name, value)
+        
+        if (name, None) in self._vector.inputs:
+            self._vector.set((name, None), value)
+        for i in range(16):
+            if (name, i) in self._vector.inputs:
+                self._vector.set((name, i), bool(value & (1 << i)))
+
+    def tick(self):
+        """Raise the common `clock` signal (and propagate state changes eagerly)."""
+        if ('common.clock', 0) in self._vector.inputs:
+            self._vector._propagate()
+            self._vector.set(('common.clock', 0), 1)  # TODO: first check that it was low
+
+    def tock(self):
+        """Lower the common `clock` signal (and propagate state changes eagerly)."""
+        if ('common.clock', 0) in self._vector.inputs:
+            self._vector._propagate()
+            self._vector.set(('common.clock', 0), 0)  # TODO: first check that it was high
+        self._vector._flop()
+
+    def ticktock(self):
+        """Raise and then lower the common `clock` signal."""
+        self.tick()
+        self.tock()
+
+    def __getattr__(self, name):
+        """Get the value of a single- or multiple-bit output."""
+
+        if (name, None) in self._vector.outputs:
+            return extend_sign(self._vector.get((name, None)))
+        else:
+            tmp = 0
+            for i in range(16):
+                if (name, i) in self._vector.outputs and self._vector.get((name, i)):
+                    tmp |= 1 << i
+            return extend_sign(tmp)
+
+    def get_internal(self, name):
+        """Get the value of a single- or multiple-bit signal which is internal to the component."""
+        if (name, None) in self._vector.internal:
+            return extend_sign(self._vector.get_internal((name, None)))
+        else:
+            tmp = 0
+            for i in range(16):
+                if (name, i) in self._vector.internal and self._vector.get_internal((name, i)):
+                    tmp |= 1 << i
+            return extend_sign(tmp)
+
+    def outputs(self):
+        return dict([(name, self.__getattr__(name)) for (name, _) in self._vector.outputs.keys()])
+
+    def internal(self):
+        return dict([(name, self.get_internal(name)) for (name, _) in self._vector.internal.keys()])
+        
+    def components(self, types):
+        """List of internal components (e.g. RAM, ROM).
+        
+        Note: types should be one or more of the types defined in nand.component, not the wrappers
+        with the same names defined in this module.
+        """
+        return [c for c in self._components if isinstance(c, types)]
+
+    def __repr__(self):
+        return str(self.outputs())
+
+
+class NandVectorComputerWrapper(NandVectorWrapper):
+    """Wrapper with extra operations for the full Computer."""
+    
+    def __init__(self, vector, stateful):
+        NandVectorWrapper.__init__(self, vector)
+        self._stateful = stateful
+
+    def run_program(self, instructions):
+        """Install and run a sequence of instructions, stopping when pc runs off the end."""
+
+        self.init_rom(instructions)
+
+        while self.pc <= len(instructions):
+            self.ticktock()
+
+    def reset_program(self):
+        """Reset pc so the program will run again from the top."""
+
+        self.reset = 1
+        self.ticktock()
+        self.reset = 0
+
+    def peek(self, address):
+        """Read a single word from the Computer's memory."""
+
+        mem, = [c for c in self._stateful if isinstance(c, RAMOps) and c.comp.address_bits == 14]
+        return mem.storage[address]
+
+    def poke(self, address, value):
+        """Write a single word to the Computer's memory."""
+
+        mem, = [c for c in self._stateful if isinstance(c, RAMOps) and c.comp.address_bits == 14]
+        mem.storage[address] = value
+
+    def init_rom(self, instructions):
+        """Overwrite the top of the ROM with a sequence of instructions.
+
+        If there's any space left over, an two-instruction infinite loop is written immediately
+        after the program, which could in theory be used to detect termination.
+        """
+
+        size = len(instructions)
+        prg = instructions + [
+            size,  # @size (which is the address of this instruction)
+            0b111_0_000000_000_111,  # JMP
+        ]
+        rom, = [c for c in self._stateful if isinstance(c, ROMOps)]
+        rom.storage = prg
