+++ test_05.py
@@ -305,7 +305,7 @@ def test_computer_no_program():
     computer = run(Computer)
     
     for _ in range(100):
-        computer.tick(); computer.tock()
+        computer.ticktock()
     
     assert computer.pc == 100
 
@@ -324,23 +324,23 @@ def test_computer_add():
     computer = run(Computer)
     
     # First run (at the beginning PC=0)
-    run_program(computer, ADD_PROGRAM)
+    computer.run_program(ADD_PROGRAM)
     
-    assert peek(computer, 0) == 5
+    assert computer.peek(0) == 5
     
 
     # Reset the PC
     computer.reset = 1
-    computer.tick(); computer.tock()
+    computer.ticktock()
     assert computer.pc == 0
     
     # Second run, to check that the PC was reset correctly.
-    poke(computer, 0, 12345)
+    computer.poke(0, 12345)
     computer.reset = 0    
     while computer.pc < len(ADD_PROGRAM):
-        computer.tick(); computer.tock()
+        computer.ticktock()
 
-    assert peek(computer, 0) == 5        
+    assert computer.peek(0) == 5        
 
 
 MAX_PROGRAM = [
@@ -365,23 +365,23 @@ MAX_PROGRAM = [
 def test_computer_max():
     computer = run(Computer)
 
-    init_rom(computer, MAX_PROGRAM)
+    computer.init_rom(MAX_PROGRAM)
 
     # first run: compute max(3,5)
-    poke(computer, 0, 3)
-    poke(computer, 1, 5)
+    computer.poke(0, 3)
+    computer.poke(1, 5)
     for _ in range(14):
-        computer.tick(); computer.tock()    
-    assert peek(computer, 2) == 5
+        computer.ticktock()    
+    assert computer.peek(2) == 5
 
     # second run: compute max(23456,12345)
-    reset_program(computer)
-    poke(computer, 0, 23456)
-    poke(computer, 1, 12345)
+    computer.reset_program()
+    computer.poke(0, 23456)
+    computer.poke(1, 12345)
     # The run on these inputs needs less cycles (different branching)
     for _ in range(10):
-        computer.tick(); computer.tock()    
-    assert peek(computer, 2) == 23456
+        computer.ticktock()    
+    assert computer.peek(2) == 23456
 
  
 def cycles_per_second():
@@ -393,17 +393,17 @@ def cycles_per_second():
     
     computer = run(Computer)
 
-    init_rom(computer, MAX_PROGRAM)
+    computer.init_rom(MAX_PROGRAM)
 
     def once():
         x = random.randint(0, 0x7FFF)
         y = random.randint(0, 0x7FFF)
-        reset_program(computer)
-        poke(computer, 0, x)
-        poke(computer, 1, y)
+        computer.reset_program()
+        computer.poke(0, x)
+        computer.poke(1, y)
         for _ in range(14):
-            computer.tick(); computer.tock()
-        assert peek(computer, 2) == max(x, y)
+            computer.ticktock()
+        assert computer.peek(2) == max(x, y)
 
     count, time = timeit.Timer(once).autorange()
 
@@ -414,66 +414,3 @@ def test_speed():
     cps = cycles_per_second()
     print(f"Measured speed: {cps:0,.1f} cycles/s")
     assert cps > 1000
- 
-
-def run_program(computer, instructions):
-    """Install and run a sequence of instructions, stopping when pc runs off the end."""
-    
-    init_rom(computer, instructions)
-
-    while computer.pc <= len(instructions):
-        computer.tick(); computer.tock()    
-
-def reset_program(computer):
-    """Reset pc so the program will run again from the top."""
-    
-    computer.reset = 1
-    computer.tick(); computer.tock()
-
-    computer.reset = 0
-    
-
-# def run_fully(computer):
-#     reset_program(computer)
-#
-#     # TODO: what is a safe termination condition? How to detect infinite loop vs other tight loop?
-#     # For example, something like this might be useful: MD=D-1; A-D; JLT
-#     while ?:
-#       computer.tick(); computer.tock()
-
-
-def peek(computer, address):
-    """Read a single word from the Computer's memory."""
-    
-    mem = get_ram(computer, address_bits=14)
-    return mem.storage[address]
-
-
-def poke(computer, address, value):
-    """Write a single word to the Computer's memory."""
-    
-    mem = get_ram(computer, address_bits=14)
-    mem.storage[address] = value
-
-
-def init_rom(computer, instructions):
-    """Overwrite the top of the ROM with a sequence of instructions.
-    
-    If there's any space left over, an two-instruction infinite loop is written immediately
-    after the program, which could in theory be used to detect termination.
-    """
-    
-    size = len(instructions)
-    prg = instructions + [
-        size,  # @size (which is the address of this instruction)
-        0b111_0_000000_000_111,  # JMP
-    ]
-    rom, = computer.components(nand.component.ROM)
-    rom.program(prg)
-
-
-def get_ram(computer, address_bits):
-    """Find one of the memories from inside the computer by matching the number of address bits.
-    """
-    # HACK!
-    return [mem for mem in computer.components(nand.component.RAM) if mem.address_bits == address_bits][0]
